add spooky

+++ 0x12-singly_linked_lists/lists.h
@@ -17,7 +17,7 @@ struct list_t *next;
 size_t print_list(const list_t *h);
 size_t list_len(const list_t *h);
 list_t *add_node(list_t **head, const char *str);
-int _strlen(char *s);
+size_t _strlen(char *s);
 list_t *add_node_end(list_t **head, const char *str);
 void free_list(list_t *head);
